+++ src/commonMain/kotlin/com/algolia/search/configuration/ConfigurationAnalytics.kt
@@ -1,9 +1,10 @@
 package com.algolia.search.configuration
 
 import com.algolia.search.client.ClientAnalytics
+import com.algolia.search.configuration.Region.Analytics.US
 import com.algolia.search.model.APIKey
 import com.algolia.search.model.ApplicationID
-import com.algolia.search.transport.analyticsHosts
+import com.algolia.search.transport.hosts
 import io.ktor.client.HttpClientConfig
 import io.ktor.client.engine.HttpClientEngine
 import io.ktor.client.features.logging.LogLevel
@@ -14,14 +15,36 @@ import io.ktor.client.features.logging.LogLevel
 public data class ConfigurationAnalytics(
     override val applicationID: ApplicationID,
     override val apiKey: APIKey,
+    val region: Region.Analytics,
     override val writeTimeout: Long = defaultWriteTimeout,
     override val readTimeout: Long = defaultReadTimeout,
     override val logLevel: LogLevel = defaultLogLevel,
-    override val hosts: List<RetryableHost> = analyticsHosts,
+    override val hosts: List<RetryableHost> = region.hosts,
     override val defaultHeaders: Map<String, String>? = null,
     override val engine: HttpClientEngine? = null,
     override val httpClientConfig: (HttpClientConfig<*>.() -> Unit)? = null
 ) : Configuration, Credentials {
+
+    @Deprecated(
+        message = "Explicitly specify an analytics region",
+        level = DeprecationLevel.WARNING,
+        replaceWith = ReplaceWith("ConfigurationAnalytics(applicationID, apiKey, Region.Analytics.US)")
+    )
+    constructor(
+        applicationID: ApplicationID,
+        apiKey: APIKey,
+        writeTimeout: Long = defaultWriteTimeout,
+        readTimeout: Long = defaultReadTimeout,
+        logLevel: LogLevel = defaultLogLevel,
+        hosts: List<RetryableHost>? = null,
+        defaultHeaders: Map<String, String>? = null,
+        engine: HttpClientEngine? = null,
+        httpClientConfig: (HttpClientConfig<*>.() -> Unit)? = null,
+        region: Region.Analytics = US
+    ) : this(
+        applicationID, apiKey, region, writeTimeout, readTimeout, logLevel,
+        hosts ?: region.hosts, defaultHeaders, engine, httpClientConfig
+    )
 
     override val compression: Compression = Compression.None
     override val httpClient = getHttpClient()