+++ CHANGELOG.md
@@ -5,6 +5,7 @@ this due the [breaking changes](https://github.com/Kotlin/kotlinx.serialization/
 
 ### Added
 - Recommendation client (#148)
+- Add analytics regions (#170)
 - Natural languages parameter to search query (#173)
 - Logs inner queries (#174)